% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ll.R
\name{calc_ll_pg_splm}
\alias{calc_ll_pg_splm}
\title{Log-likelihood for pg_splm() model for use in model selection}
\usage{
calc_ll_pg_splm(Y, X, out)
}
\arguments{
\item{Y}{is a \eqn{n \times J}{n x J} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of climate variables.}

\item{out}{is a list of MCMC outputs from \code{pg_splm()}}
}
\description{
this function generates the log-linkelihood for data fit using the pg_splm()
function. The log-likelihood can be used for model selection and evaluation.
}
