% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ll.R
\name{calc_ll_pg_stlm_loo}
\alias{calc_ll_pg_stlm_loo}
\title{Leave-one-out log-likelihood for pg_stlm() model for use in model selection}
\usage{
calc_ll_pg_stlm_loo(Y, X, locs, out, n_message = 50)
}
\arguments{
\item{Y}{is a \eqn{N \times J \times T}{N x J x T} array of compositional count data.}

\item{X}{is a \eqn{N \times p}{n_sites x p} matrix of climate variables.}

\item{locs}{is a \eqn{N \times 2}{N x 2} matrix of observation locations}

\item{out}{is a list of MCMC outputs from \code{pg_stlm()}, \code{pg_stlm_overdispersed}, or \code{pg_stlm_latent_overdispersed}}

\item{n_message}{A positive integer number of frequency of iterations
to output a progress message. For example, \code{n_message = 50}
outputs progress messages every 50 iterations.}
}
\description{
this function generates the log-likelihood for data fit using the pg_splm()
function. The log-likelihood can be used for model selection and evaluation.
}
