% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_Mi}
\alias{calc_Mi}
\title{Calculate the intermediate Polya-gamma value Mi}
\usage{
calc_Mi(Y)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of compositional count data.}
}
\description{
Calculate the intermediate Polya-gamma value Mi
}
