% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inits.R
\name{check_inits_pg_splm}
\alias{check_inits_pg_splm}
\title{Check initial values}
\usage{
check_inits_pg_splm(Y, X, locs, inits)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of multinomial count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of observation locations.}

\item{inits}{is the list of current initial values}
}
\description{
this function check that the initial values for \code{pg_splm()} are properly specified
}
\keyword{internal}
