% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inits.R
\name{check_inits_pg_stlm}
\alias{check_inits_pg_stlm}
\title{Check initial values}
\usage{
check_inits_pg_stlm(
  Y,
  X,
  locs,
  inits,
  corr_fun = "exponential",
  shared_covariance_params = TRUE
)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of multinomial count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of observation locations.}

\item{inits}{is the list of current initial values}

\item{corr_fun}{is a character that denotes the correlation function form. Current options include "matern" and "exponential".}

\item{shared_covariance_params}{is a logical input that determines whether to fit the spatial process with component specifice parameters. If TRUE, each component has conditionally independent Gaussian process parameters theta and tau2. If FALSE, all components share the same Gaussian process parameters theta and tau2.}
}
\description{
this function check that the initial values for \code{pg_stlm()} are properly specified
}
\keyword{internal}
