% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inits.R
\name{check_inits_pgSPLM}
\alias{check_inits_pgSPLM}
\title{Check initial values}
\usage{
check_inits_pgSPLM(Y, X, inits)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of multinomial count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{inits}{is the list of current initial values}
}
\description{
this function check that the initial values for pgSPLM are properly specified
}
\keyword{internal}
