% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inits.R
\name{check_inits_pgSTLM}
\alias{check_inits_pgSTLM}
\title{Check initial values}
\usage{
check_inits_pgSTLM(Y, X, inits)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of multinomial count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{inits}{is the list of current initial values}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of observation locations.}

\item{corr_fun}{is a character that denotes the correlation function form. Current options include "matern" and "exponential".}
}
\description{
this function check that the initial values for pgSTLM are properly specified
}
\keyword{internal}
