% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input.R
\name{check_input_pg_lm}
\alias{check_input_pg_lm}
\title{Check inputs}
\usage{
check_input_pg_lm(Y, X)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of climate variables.}
}
\description{
this function check that the input values are properly specified
}
\keyword{internal}
