% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-spatial.R
\name{check_input_spatial}
\alias{check_input_spatial}
\title{Check inputs}
\usage{
check_input_spatial(Y, X, locs)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of climate variables.}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of spatial locations.}
}
\description{
this function check that the input values are properly specified
}
