% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-params.R
\name{check_params}
\alias{check_params}
\title{Check parameters}
\usage{
check_params(params)
}
\arguments{
\item{params}{is the list of current parameter settings}
}
\description{
this function check that the parameter values are properly specified
}
