% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-priors.R
\name{check_priors}
\alias{check_priors}
\title{Check priors}
\usage{
check_priors(params, priors)
}
\arguments{
\item{params}{is the list of current parameter settings}

\item{priors}{is the list of current prior values}
}
\description{
this function check that the prior values are properly specified
}
\keyword{internal}
