% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-priors.R
\name{check_priors_pg_lm}
\alias{check_priors_pg_lm}
\title{Check priors}
\usage{
check_priors_pg_lm(Y, X, priors)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{priors}{is the list of current prior values}
}
\description{
this function check that the prior values are properly specified for the \code{pg_lm()} model
}
\keyword{internal}
