% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-priors.R
\name{check_priors_pg_splm}
\alias{check_priors_pg_splm}
\title{Check priors}
\usage{
check_priors_pg_splm(Y, X, priors, corr_fun = "exponential")
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{priors}{is the list of current prior values}

\item{corr_fun}{is the correlation function}
}
\description{
this function check that the prior values are properly specified for the \code{pg_splm()} model.
}
\keyword{internal}
