% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-to-coda.R
\name{convert_to_coda}
\alias{convert_to_coda}
\title{Convert a list of MCMC output from the pgR package into coda format}
\usage{
convert_to_coda(out)
}
\arguments{
\item{out}{is a list of MCMC output where each element of the list
is an MCMC chain}
}
\value{
fill out later XXX
}
\description{
This function  process custom MCMC output from Rcpp into an R coda
format. The function assumes the first dimension of each parameter
within the list (a MCMC chain) is the number of MCMC iterations
}
\keyword{internal}
