% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-function.R
\name{correlation_function}
\alias{correlation_function}
\title{Generate Matern covariance matrix}
\usage{
correlation_function(D, theta)
}
\arguments{
\item{D}{is an \eqn{n \times n}{n x n} matrix of pairwise distances between locations}

\item{theta}{is a \eqn{2}{2}-dimensional vector of Matern covariance parameter inputs on the log-scale with the first element the Matern smoothness parameter \code{nu} and second element the Matern \code{range} parameter}
}
\description{
A function for generatingg the Matern covariance matrix
}
