% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-to-propotions.R
\name{counts_to_proportions}
\alias{counts_to_proportions}
\title{Convert compsitional count data to proportions}
\usage{
counts_to_proportions(Y)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of compositional count data where each of the \eqn{n}{n} rows represents a compositional count observation.}
}
\description{
this function takes a matrix of compositional count data and converts the data to a proportion
}
