% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-inits.R
\name{default_inits_pg_lm}
\alias{default_inits_pg_lm}
\title{Initialized default initial conditions}
\usage{
default_inits_pg_lm(Y, X, priors)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of multinomial count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{priors}{is a list of prior settings.}
}
\description{
A function for setting up the default \code{pg_lm()} initial values
}
