% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-inits.R
\name{default_inits_pg_splm}
\alias{default_inits_pg_splm}
\title{Initialized default initial conditions}
\usage{
default_inits_pg_splm(
  Y,
  X,
  priors,
  corr_fun = "exponential",
  shared_covariance_params = TRUE
)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of multinomial count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{priors}{is a list of prior settings.}

\item{corr_fun}{is a character that denotes the correlation function form. Current options include "matern" and "exponential".}

\item{shared_covariance_params}{is a logical input that determines whether to fit the spatial process with component specifice parameters. If TRUE, each component has conditionally independent Gaussian process parameters theta and tau2. If FALSE, all components share the same Gaussian process parameters theta and tau2.}
}
\description{
A function for setting up the default \code{pg_splm()} initial values
}
