% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-inits.R
\name{default_inits_pgLM}
\alias{default_inits_pgLM}
\title{Initialized default intial conditions}
\usage{
default_inits_pgLM(Y, X, priors)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of multinomial count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of variables.}

\item{priors}{is a list of prior settings.}
}
\description{
A function for setting up the default pgLM priors
}
