% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-priors.R
\name{default_priors_pg_stlm_mra}
\alias{default_priors_pg_stlm_mra}
\title{Initialized default values of priors for \code{pg_stlm_mra()}}
\usage{
default_priors_pg_stlm_mra(Y, X)
}
\arguments{
\item{Y}{is a \eqn{n \times d \times T}{n x d x T} array of compositional count data through time.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of climate variables.}
}
\description{
A function for setting up the default \code{pg_stlm_mra()} priors
}
