% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-priors.R
\name{default_priors_pgLM}
\alias{default_priors_pgLM}
\title{Initialized default values of priors}
\usage{
default_priors_pgLM(Y, X)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of climate variables.}
}
\description{
A function for setting up the default pgLM priors
}
