% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnrm_arma_mc}
\alias{dmvnrm_arma_mc}
\title{A function for evaluating a multivariate normal density in parallel}
\usage{
dmvnrm_arma_mc(x, mean, sigma, logd = FALSE, cores = 1L)
}
\arguments{
\item{x}{An \eqn{N \times d}{N x d} \code{matrix} of multivariate Gaussian observations.}

\item{mean}{\code{mean} A \eqn{d} \code{vector} of the Gaussian distribution mean}

\item{sigma}{A \eqn{d \times d}{d x d} positive definite covariance \code{matrix}}

\item{logd}{A logical value indicating whether the funtion should return the density on the log scale}

\item{cores}{An integer that gives the number of cores for openMP parallelization}
}
\description{
A function for evaluating a multivariate normal density in parallel
}
