% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eta_to_pi}
\alias{eta_to_pi}
\title{Convert from eta to pi}
\usage{
eta_to_pi(eta)
}
\arguments{
\item{eta}{The latent intensity vector to be transformed into a probability}
}
\description{
Convert from eta to pi
}
