% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_dist_matrix}
\alias{is_dist_matrix}
\title{Check if a distance matrix of dimension \eqn{n \times n}{n x n} (symmetic and non-negative values)}
\usage{
is_dist_matrix(x, n, m)
}
\arguments{
\item{x}{is the input}

\item{n}{is the input matrix dimension (assuming a square matrix)}
}
\description{
this function checks if the input is a symmetrix positive definite matrix
}
\keyword{internal}
