% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_positive_numeric_matrix}
\alias{is_positive_numeric_matrix}
\title{Check if a positive numeric matrix of dimension \eqn{n \times m}{n x m}}
\usage{
is_positive_numeric_matrix(x, n, m)
}
\arguments{
\item{x}{is the input}

\item{n}{is the input length}
}
\description{
this function checks if the input is a positive numeric (double) matrix of dimension \eqn{n \times m}{n x m}
}
\keyword{internal}
