% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_alpha.R
\name{make_alpha}
\alias{make_alpha}
\title{Contruct the Dirichlet-multinomial parameter matrix \code{alpha}}
\usage{
make_alpha(Xbs, beta, params)
}
\arguments{
\item{Xbs}{is an \eqn{n \times q} matrix that are inputs into the functional form}

\item{beta}{is an \eqn{q} vector of inputs into the functional form}

\item{params}{is a list of parameters for each functional type and link function. See more details for details}
}
\description{
A function for calculating the Dirichlet-multinomial parameters from the underlying latent parameters
}
