% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gelman_rubin.R
\name{make_gelman_rubin}
\alias{make_gelman_rubin}
\title{Calculate the Gelman-Rubin statistic for custom MCMC output}
\usage{
make_gelman_rubin(out)
}
\arguments{
\item{\code{out}}{is a list object where each element of
the list is the results from a single MCMC chain. Within
each MCMC chain, there will be a list of objects of various
sizes where the first dimension of each of these objects 
is the number of MCMC iterations.}
}
\description{
A function to calculate Gelman-Rubin R-hat statistics for 
custom coded model MCMC output=
}
