% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{MCMC wrapper for models in package \code{pollenReconstruction}}
\usage{
mcmc(Y, X, params, priors, inits = NULL, config = NULL)
}
\arguments{
\item{Y}{is a \eqn{n \times d}{n x d} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of climate variables.}

\item{params}{is the list of parameter settings.}

\item{priors}{is the list of prior settings.}

\item{inits}{is the list of intial values if the user wishes to specify initial values. If these values are not specified, then the intital values will be randomly sampled from the prior.}

\item{config}{is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.}
}
\description{
this function setups the model and runs the desired MCMC model.
}
