% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra-wendland-2d-pred.R
\name{mra_wendland_2d_pred}
\alias{mra_wendland_2d_pred}
\title{Code to construct the mutli-resolution sparse basis function representation for fitting spatial processes}
\usage{
mra_wendland_2d_pred(
  locs,
  locs_pred,
  M = 4,
  n_neighbors = 25,
  padding = 0.1,
  use_spam = TRUE
)
}
\arguments{
\item{locs}{The location variables in 2 dimensions over which to construct the basis function representation}

\item{M}{The number of resolutions}

\item{n_neighbors}{}

\item{padding}{}

\item{use_spam}{}
}
\value{

}
\description{
Code to construct the mutli-resolution sparse basis function representation for fitting spatial processes
}
