% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_connection.R
\name{pg_dplyr}
\alias{pg_dplyr}
\title{Connection to postgresql via dplyr. The password is entered by .rs.askForPassword().}
\usage{
pg_dplyr(pg_src, schema = public, table)
}
\arguments{
\item{pg_src:}{postgresql connection.}

\item{schema:}{schema name. Default(public).}

\item{table:}{table name.}
}
\description{
Connection to postgresql via dplyr. The password is entered by .rs.askForPassword().
}

