% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_connection.R
\name{pg_dplyr_tbl}
\alias{pg_dplyr_tbl}
\title{Read table from postgresql via dplyr. The password is entered by .rs.askForPassword().}
\usage{
pg_dplyr_tbl(dbname = NULL, schema = public, table = NULL,
  host = localhost, port = 5432, user = postgres,
  password = .rs.askForPassword(paste("Password for user", user)))
}
\arguments{
\item{dbname:}{database name.}

\item{user:}{user name. Default(postgres).}

\item{host:}{host. Default(localhost).}

\item{port:}{port. Default(5432).}

\item{password:}{password. Default(.rs.askForPassword()).}

\item{schema:}{schema. Default(public).}

\item{table:}{table. Default(NULL)}
}
\description{
Read table from postgresql via dplyr. The password is entered by .rs.askForPassword().
}

