% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_funs.R
\name{pg_export}
\alias{pg_export}
\title{Export query to CSV with delimiter tabular and header.}
\usage{
pg_export(
  con,
  query,
  dir = file.choose(new = T),
  encoding = "WIN1252",
  delim = "\\t"
)
}
\arguments{
\item{con:}{connection.}

\item{query:}{(chr) SQL query.}

\item{dir:}{(chr) directory. Default(file.choose()).}

\item{encoding:}{(chr) SET client_encoding to. Default 'WIN1252'.}

\item{delim:}{(chr) Set delimiter. Default tabular.}
}
\description{
Export query to CSV with delimiter tabular and header.
}
