% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_funs.R
\name{pg_importdbf}
\alias{pg_importdbf}
\title{Read tabular data in dbf file and write it in PostgreSQL.}
\usage{
pg_importdbf(file = file.choose(), con, schema = "public", table_name)
}
\arguments{
\item{file:}{path excel file.}

\item{con:}{Pg connection.}

\item{schema:}{name of schema to write.}

\item{table_name:}{table name.}
}
\description{
Read tabular data in dbf file and write it in PostgreSQL.
}
