% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-mvgp.R
\name{pg_mvgp}
\alias{pg_mvgp}
\title{Bayesian Polya-gamma regression}
\usage{
pg_mvgp(
  Y,
  X,
  Z,
  locs,
  site_idx,
  year_idx,
  params,
  priors,
  n_cores = 1L,
  inits = NULL,
  config = NULL,
  n_chain = 1
)
}
\arguments{
\item{Y}{is a \eqn{n \times J}{n x J} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of fixed effects (like latitude, elevation, etc)}

\item{Z}{is Z}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of observation locations.}

\item{site_idx}{is site_idx}

\item{year_idx}{is year_idx}

\item{params}{is the list of parameter settings.}

\item{priors}{is the list of prior settings.}

\item{n_cores}{is the number of cores for parallel computation using openMP.}

\item{inits}{is the list of intial values if the user wishes to specify initial values. If these values are not specified, then the intital values will be randomly sampled from the prior.}

\item{config}{is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.}

\item{n_chain}{is the MCMC chain id. The default is 1.}
}
\description{
this function runs the Bayesian multinomial regression using Polya-gamma data augmentation
}
