% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_funs.R
\name{pg_nomcols}
\alias{pg_nomcols}
\title{column name and data type from tbl.}
\usage{
pg_nomcols(con, tbl)
}
\arguments{
\item{con:}{Conection.}

\item{tbl:}{chr table name.}
}
\description{
column name and data type from tbl.
}
