% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_funs.R
\name{pg_readb}
\alias{pg_readb}
\title{Connect to several tables from db using dplyr}
\usage{
pg_readb(db = NULL, schema = "public", tbls = NULL, host = "localhost")
}
\arguments{
\item{db}{chr database name}

\item{schema}{chr schema name. Default 'public'.}

\item{tbls}{chr or vector of chrs with table names.}

\item{host}{chr. Defaulta 'localhost'.}
}
\description{
Connect to several tables from db using dplyr
}
