% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_funs.R
\name{pg_save}
\alias{pg_save}
\title{Save data.frame into Postgresql}
\usage{
pg_save(df, con, schema, table_name, overwrite = FALSE, append = FALSE)
}
\arguments{
\item{con:}{pg connection.}

\item{schema:}{schema name.}

\item{table_name:}{name that will have table.}

\item{df:}{data.frame to save.}
}
\description{
Save data.frame into Postgresql
}

