% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_funs.R
\name{pg_schema}
\alias{pg_schema}
\title{Summary of schema. Additionaly if use first and second arguments read table or view.}
\usage{
pg_schema(con, schema, first = NULL, second = NULL)
}
\arguments{
\item{con:}{pg connection.}

\item{schema:}{schema.}

\item{first:}{int 1 = Tables. 2 = Views.}

\item{second:}{int with table or view position.}

\item{to_clip:}{logical if TRUE copies to clipboard}
}
\description{
Summary of schema. Additionaly if use first and second arguments read table or view.
}
