% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_funs.R
\name{pg_schemas}
\alias{pg_schemas}
\title{Show schemas information from database}
\usage{
pg_schemas(con)
}
\arguments{
\item{con:}{DBI connection.}
}
\description{
Show schemas information from database
}
