% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_funs.R
\name{pg_show}
\alias{pg_show}
\title{Show information about, schema and table depend on arguments presents.}
\usage{
pg_show(con, schema = NULL, table = NULL)
}
\arguments{
\item{con:}{DBI connection.}

\item{schema:}{chr. Schema name.}

\item{table:}{chr. Table name.}
}
\description{
Show information about, schema and table depend on arguments presents.
}
