% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_splm_mra.R
\name{pg_splm_mra}
\alias{pg_splm_mra}
\title{Bayesian Polya-gamma regression}
\usage{
pg_splm_mra(
  Y,
  X,
  locs,
  params,
  priors,
  n_cores = 1L,
  M = 4,
  n_coarse_grid = 10,
  inits = NULL,
  config = NULL,
  n_chain = 1,
  progress = FALSE,
  verbose = FALSE,
  use_spam = TRUE
)
}
\arguments{
\item{Y}{is a \eqn{n \times J}{n x J} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of climate variables.}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of observation locations.}

\item{params}{is a list of parameter settings. The list
\code{params} must contain the following values:
\itemize{
\item \code{n_adapt}: A positive integer number of adaptive MCMC iterations.
\item \code{n_mcmc}: A positive integer number of total MCMC iterations
post adaptation.
\item \code{n_thin}: A positive integer number of MCMC iterations per saved
sample.
\item \code{n_message}: A positive integer number of frequency of iterations
to output a progress message. For example, \code{n_message = 50}
outputs progress messages every 50 iterations.
}}

\item{priors}{is the list of prior settings.}

\item{n_cores}{is the number of cores for parallel computation using openMP.}

\item{M}{The number of resolutions.}

\item{n_coarse_grid}{The number of basis functions in one direction (e.g. \code{n_coarse_grid = 10} results in a \eqn{10 \times 10}{10x10} course grid which is further extended by the number of additional padding basis functions given by \code{n_padding}.}

\item{inits}{is the list of initial values if the user wishes to specify initial values. If these values are not specified, then the intital values will be randomly sampled from the prior.}

\item{config}{is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.}

\item{n_chain}{is the MCMC chain id. The default is 1.}

\item{progress}{is a logical input that determines whether to print a progress bar.}

\item{verbose}{is a logical input that determines whether to print more detailed messages.}

\item{use_spam}{is a boolean flag to determine whether the output is a list of spam matrix objects (\code{use_spam = TRUE}) or a an \eqn{n \times n}{n x n} sparse Matrix of class "dgCMatrix" \code{use_spam = FALSE} (see spam and Matrix packages for details).}
}
\description{
this function runs the Bayesian multinomial regression using Polya-gamma data augmentation
}
