% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_funs.R
\name{pg_sql}
\alias{pg_sql}
\title{Send query and return data.frame.}
\usage{
pg_sql(con, query)
}
\arguments{
\item{con:}{conection.}

\item{query:}{sql query.}
}
\description{
Send query and return data.frame.
}
