% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_connection.R
\name{pg_src}
\alias{pg_src}
\title{Connection to postgresql via dplyr. The password is entered by .rs.askForPassword().}
\usage{
pg_src(dbname = NULL, host = localhost, port = 5432, user = postgres)
}
\arguments{
\item{dbname:}{database name.}

\item{user:}{user name. Default(postgres).}

\item{host:}{host. Default(localhost).}
}
\description{
Connection to postgresql via dplyr. The password is entered by .rs.askForPassword().
}

