% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_funs.R
\name{pg_tables}
\alias{pg_tables}
\title{Table information from schema}
\usage{
pg_tables(con, schema)
}
\arguments{
\item{con:}{DBI conection.}

\item{schema:}{chr. schema name.}
}
\description{
Table information from schema
}
