% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_funs.R
\name{pg_tbl}
\alias{pg_tbl}
\title{Summarise information about a tbl}
\usage{
pg_tbl(con, tbl)
}
\arguments{
\item{con:}{DBI connection.}

\item{tbl:}{chr. tbl name.}
}
\description{
Summarise information about a tbl
}
