% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgCon.R
\name{PgCon}
\alias{PgCon}
\title{R6 Class PyCon postgresql connection.}
\description{
R6 Class PyCon postgresql connection.

R6 Class PyCon postgresql connection.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PgCon-new}{\code{PgCon$new()}}
\item \href{#method-PgCon-finalize}{\code{PgCon$finalize()}}
\item \href{#method-PgCon-print}{\code{PgCon$print()}}
\item \href{#method-PgCon-import}{\code{PgCon$import()}}
\item \href{#method-PgCon-export}{\code{PgCon$export()}}
\item \href{#method-PgCon-disconnect}{\code{PgCon$disconnect()}}
\item \href{#method-PgCon-describeTable}{\code{PgCon$describeTable()}}
\item \href{#method-PgCon-clone}{\code{PgCon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PgCon-new"></a>}}
\if{latex}{\out{\hypertarget{method-PgCon-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PgCon$new(
  name,
  user = "postgres",
  host = "localhost",
  port = "5432",
  driver = RPostgres::Postgres(),
  client_encoding = "UTF8"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PgCon-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-PgCon-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PgCon$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PgCon-print"></a>}}
\if{latex}{\out{\hypertarget{method-PgCon-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PgCon$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PgCon-import"></a>}}
\if{latex}{\out{\hypertarget{method-PgCon-import}{}}}
\subsection{Method \code{import()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PgCon$import(sql_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PgCon-export"></a>}}
\if{latex}{\out{\hypertarget{method-PgCon-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PgCon$export(
  df,
  schema = "public",
  table_name,
  overwrite = FALSE,
  append = FALSE,
  temporary = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PgCon-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-PgCon-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PgCon$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PgCon-describeTable"></a>}}
\if{latex}{\out{\hypertarget{method-PgCon-describeTable}{}}}
\subsection{Method \code{describeTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PgCon$describeTable(table_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PgCon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PgCon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PgCon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
