\name{pgR-package}
\alias{pgR-package}
\alias{pgR}
\docType{package}
\title{
\packageTitle{pgR}
}
\description{
\packageDescription{pgR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pgR}
\packageIndices{pgR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pgR}

Maintainer: \packageMaintainer{pgR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
