% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{pgR}
\alias{pgR}
\title{pgR}
\description{
Fit Bayesian Polya-gamma regression models. This package contains code to fit multinomial regression, multinomial spatial regression, mutlinomial spatially varying coefficient models, sparse multiresolution multinomial spatial models, and ...
}
\author{
John Tipton
}
\keyword{internal}
