% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-function.R
\name{plot_correlation_function}
\alias{plot_correlation_function}
\title{Plot the correlation function}
\usage{
plot_correlation_function(
  d,
  theta,
  corr_fun = "exponential",
  base_size = 12,
  file = NULL,
  width = 16,
  height = 9
)
}
\arguments{
\item{d}{is an \eqn{n \times 1}{n x 1} vector of distances}

\item{theta}{is the correlation function parameters. If
\code{corr_fun} is "matern", then \code{theta} is a
\eqn{2}{2}-dimensional vector of Matern covariance parameter inputs
on the log-scale with the first element the Matern range parameter
\code{range} on the log-scale and second element the Matern smoothness parameter
\code{nu} on the log-scale. If \code{corr_fun} is "exponential" then \code{theta} is
a positive number representing the exponential correlation function
range parameter on the log-scale.}

\item{corr_fun}{is the correlation function}

\item{base_size}{The base size for the plot}

\item{file}{If \code{file = NULL}, the ggplot object is returned. If \code{file} is not NULL, an image is saved to the file path specified by \code{file}}

\item{width}{If a file path is specified, \code{width} determines the width of the saved image (in inches)}

\item{height}{If a file path is specified, \code{height} determines the height of the saved image (in inches)}
}
\value{
Either a ggplot object of the model trace plots (if \code{file = NULL}) or a saved image file with no return (\code{file} is not NULL)
}
\description{
Plot the correlation function
}
