% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-pg_lm.R
\name{predict_pg_lm}
\alias{predict_pg_lm}
\title{Bayesian Polya-gamma regression prediction}
\usage{
predict_pg_lm(out, X_pred)
}
\arguments{
\item{out}{is a list of MCMC outputs from \code{pg_lm()}}

\item{X_pred}{is a \eqn{n_{pred} \times p}{n_{pred} x p} matrix of climate variables.}
}
\description{
this function generates predictions from the Bayesian multinomial regression using Polya-gamma data augmentation
}
