% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-pg_stlm_mra.R
\name{predict_pg_stlm_mra}
\alias{predict_pg_stlm_mra}
\title{Bayesian Polya-gamma regression prediction}
\usage{
predict_pg_stlm_mra(
  out,
  X,
  X_pred,
  locs,
  locs_pred,
  corr_fun,
  shared_covariance_params,
  progress = TRUE,
  verbose = FALSE,
  force = FALSE
)
}
\arguments{
\item{out}{is a list of MCMC outputs from \code{pg_stlm()}}

\item{X}{is a \eqn{n \times p}{n x p} matrix of covariates at the observed locations.}

\item{X_pred}{is a \eqn{n_{pred} \times p}{n_{pred} x p} matrix of covariates at the locations where predictions are to be made.}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of locations where observations were taken.}

\item{locs_pred}{is a \eqn{n_pred \times 2}{n_pred x 2} matrix of locations where predictions are to be made.}

\item{corr_fun}{is a character that denotes the correlation function form. Current options include "matern" and "exponential".}

\item{shared_covariance_params}{is a logicial input that determines whether to fit the spatial process with component specifice parameters. If TRUE, each component has conditionally independent Gaussian process parameters theta and tau2. If FALSE, all components share the same Gaussian process parameters theta and tau2.}

\item{progress}{is a logicial input that determines whether to print a progress bar.}

\item{verbose}{is a logicial input that determines whether to print more detailed messages.}

\item{force}{is a logicial input that determines whether to allow for predictions at more locations than 10000. The default is FALSE}
}
\description{
this function generates predictions from the Bayesian multinomial regression using Polya-gamma data augmentation
}
