% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-pgLM.R
\name{predict_pgLM}
\alias{predict_pgLM}
\title{Bayesian Polya-gamma regression prediction}
\usage{
predict_pgLM(out, X_pred, n_cores = 1L)
}
\arguments{
\item{out}{is a list of MCMC outputs from pgLM}

\item{X_pred}{is a \eqn{n_{pred} \times p}{n_{pred} x p} matrix of climate variables.}
}
\description{
this function generates predictions from the Bayesian multinomial regression using Polya-gamma data augmentation
}
