% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_pgdraw}
\alias{rcpp_pgdraw}
\title{A function for evaluating drawing random Polya-gamma random variables in parallel}
\arguments{
\item{b}{An \eqn{N}{N} \code{vector} of Polya-gamma parameters}

\item{c}{An \eqn{N}{N} \code{vector} of Polya-gamma parameters}

\item{cores}{An integer that gives the number of cores for openMP parallelization}
}
\description{
A function for evaluating drawing random Polya-gamma random variables in parallel
}
\keyword{internal}
