% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_splines.R
\name{setup_splines}
\alias{setup_splines}
\title{this function setups the bspline parameters}
\usage{
setup_splines(params)
}
\arguments{
\item{params}{is the list of current parameter settings}
}
\description{
this function setups the bspline parameters
}
