% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{Softmax transformation}
\usage{
softmax(x)
}
\arguments{
\item{x}{is the input}
}
\description{
this function exaluates a softmax transformation of a vector on the real number line to the simplex using the log sum of exponentials function
}
