% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_connection.R
\name{pg_con}
\alias{pg_con}
\title{Connection to postgresql via RPostgreSQL. The password is entered by .rs.askForPassword().}
\usage{
pg_con(dbname = NULL, user = postgres, host = localhost)
}
\arguments{
\item{dbname:}{database name.}

\item{user:}{user name. Default(postgres).}

\item{host:}{host. Default(localhost).}
}
\description{
Connection to postgresql via RPostgreSQL. The password is entered by .rs.askForPassword().
}

