% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgCon.R
\docType{data}
\name{PgCon}
\alias{PgCon}
\title{R6 Class PyCon postgresql connection.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PgCon
}
\description{
R6 Class PyCon postgresql connection.
}
\keyword{datasets}
