% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgLM.R
\name{pgLM}
\alias{pgLM}
\title{Bayesian Polya-gamma regression}
\usage{
pgLM(
  Y,
  X,
  params,
  priors,
  n_cores = 1L,
  inits = NULL,
  config = NULL,
  n_chain = 1,
  sample_rmvn = FALSE
)
}
\arguments{
\item{Y}{is a \eqn{n \times J}{n x J} matrix of compositional count data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of climate variables.}

\item{params}{is a list of parameter settings. The list
\code{params} must contain the following values:
\itemize{
\item \code{n_adapt}: A positive integer number of adaptive MCMC iterations.
\item \code{n_mcmc}: A positive integer number of total MCMC iterations
post adaptation.
\item \code{n_thin}: A positive integer number of MCMC iterations per saved
sample.
\item \code{n_message}: A positive integer number of frequency of iterations
to output a progress message. For example, \code{n_message = 50}
outputs progress messages every 50 iterations.
}}

\item{priors}{is the list of prior settings.}

\item{n_cores}{is the number of cores for parallel computation using openMP.}

\item{inits}{is the list of intial values if the user wishes to specify initial values. If these values are not specified, then the intital values will be randomly sampled from the prior.}

\item{config}{is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.}

\item{n_chain}{is the MCMC chain id. The default is 1.}

\item{sample_rmvn}{is an indicator whether the faster multivariate normal sampler is used.}
}
\description{
this function runs the Bayesian multinomial regression using Polya-gamma data augmentation
}
